+++ Preventanyl/app/src/main/java/noverdose/preventanyl/OverdoseDialog.java
@@ -1,0 +1,48 @@
+package noverdose.preventanyl;
+
+/**
+ * Created by yudhvirraj on 2017-11-28.
+ */
+
+import android.app.AlertDialog;
+import android.app.Dialog;
+import android.content.DialogInterface;
+import android.os.Bundle;
+import android.support.v4.app.DialogFragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+/**
+ * Created by yudhvirraj on 2017-09-07.
+ */
+
+
+public class OverdoseDialog extends DialogFragment {
+
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        builder.setTitle("Overdose")
+                .setMessage("5")
+                .setCancelable(false)
+                .setPositiveButton("OKAY", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int id) {
+
+                    }
+                })
+                .setNegativeButton("Dismiss", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int id) {
+                        System.exit(0);
+                    }
+                });
+        return builder.create();
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        getDialog ().setCanceledOnTouchOutside (false);
+        return super.onCreateView(inflater, container, savedInstanceState);
+    }
+
+}